+++ yjl-service/src/main/java/com/yjl/service/ArticalService.java
@@ -1,11 +1,12 @@
 package com.yjl.service;
 
+
 import com.yjl.entity.Artical;
 
 import java.util.List;
 
 public interface ArticalService {
-    List<Artical> getByExample(int id);
+    List<Artical> getByExample(Long id);
 
-    Artical getById(int id);
+    Artical getById(Long id);
 }
